requirements.txt
streamlit>=1.35.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
